refactor(antd): refactoring FormButtonGroup typings
