+++ packages/antd/src/components/formButtonGroup.tsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
 import { Row, Col } from './grid'
 import { FormLayoutConsumer } from '../form'
+import { IFormButtonGroupProps } from '../type'
 import Sticky from 'react-stikky'
 import cls from 'classnames'
 import styled from 'styled-components'
@@ -61,20 +62,7 @@ const isElementInViewport = (
   )
 }
 
-export interface IFormButtonGroupProps {
-  sticky?: boolean
-  style?: React.CSSProperties
-  itemStyle?: React.CSSProperties
-  className?: string
-
-  triggerDistance?: any
-  offsetDistance?: any
-  zIndex?: number
-  span?: number
-  offset?: number
-}
-
-export const FormButtonGroup = styled(
+export const FormButtonGroup: React.FC<IFormButtonGroupProps> = styled(
   class FormButtonGroup extends Component<IFormButtonGroupProps> {
     public static defaultProps = {
       span: 24
